+++ packages/dart_firebase_admin/lib/app_check.dart
@@ -1,0 +1,2 @@
+export 'src/app_check/app_check.dart';
+export 'src/app_check/app_check_api.dart';