+++ packages/dart_firebase_admin/lib/security_rules.dart
@@ -1,0 +1,4 @@
+export 'src/security_rules/security_rules.dart';
+export 'src/security_rules/security_rules_api_internals.dart'
+    hide SecurityRulesApiClient;
+export 'src/security_rules/security_rules_internals.dart';