+++ packages/dart_firebase_admin/lib/src/auth.dart
@@ -3,11 +3,11 @@ import 'dart:typed_data';
 
 import 'package:collection/collection.dart';
 import 'package:dart_jsonwebtoken/dart_jsonwebtoken.dart' as dart_jsonwebtoken;
-import 'package:firebaseapis/identitytoolkit/v1.dart' as auth1;
-import 'package:firebaseapis/identitytoolkit/v1.dart' as v1;
-import 'package:firebaseapis/identitytoolkit/v2.dart' as auth2;
-import 'package:firebaseapis/identitytoolkit/v2.dart' as v2;
-import 'package:firebaseapis/identitytoolkit/v3.dart' as auth3;
+import 'package:googleapis/identitytoolkit/v1.dart' as auth1;
+import 'package:googleapis/identitytoolkit/v1.dart' as v1;
+import 'package:googleapis/identitytoolkit/v2.dart' as auth2;
+import 'package:googleapis/identitytoolkit/v2.dart' as v2;
+import 'package:googleapis/identitytoolkit/v3.dart' as auth3;
 import 'package:http/http.dart';
 import 'package:meta/meta.dart';