+++ packages/dart_firebase_admin/test/google_cloud_firestore/collection_test.dart
@@ -7,7 +7,21 @@ void main() {
   group('Collection interface', () {
     late Firestore firestore;
 
-    setUp(() => firestore = createFirestore());
+    setUp(() async => firestore = await createFirestore());
+
+    test('supports + in collection name', () async {
+      final a = firestore
+          .collection('/collection+a/lF1kvtRAYMqmdInT7iJK/subcollection');
+
+      expect(a.path, 'collection+a/lF1kvtRAYMqmdInT7iJK/subcollection');
+
+      await a.add({'foo': 'bar'});
+
+      final results = await a.get();
+
+      expect(results.docs.length, 1);
+      expect(results.docs.first.data(), {'foo': 'bar'});
+    });
 
     test('has doc() method', () {
       final collection = firestore.collection('colId');
@@ -73,6 +87,7 @@ void main() {
 
       final documentSnapshot = await documentRef.get();
 
+      expect(documentSnapshot.exists, isTrue);
       expect(documentSnapshot.data(), {'foo': 'bar'});
     });