+++ packages/dart_firebase_admin/test/mock.dart
@@ -1,5 +1,5 @@
 import 'package:dart_firebase_admin/dart_firebase_admin.dart';
-import 'package:firebaseapis/fcm/v1.dart';
+import 'package:googleapis/fcm/v1.dart';
 import 'package:http/http.dart';
 import 'package:mocktail/mocktail.dart';